requirements.txt
fastapi
gunicorn
uvicorn
opencv-python-headless
google-cloud-storage
requests
Pillow
numpy